reuse meta-tx signature methods for dispute mutual agreement
